Add helpers for accessing court metadata
This adds a courts object which allows access to metadata about the courts whose judgments we hold, loaded from the court_names.yml file in the data folder. This file has been augmented with metadata used elsewhere in the app, with the aim of it being a single source of truth for information about courts used in the app.