Add accessors for all listable courts and tribunals
This was an oversight on my part when i first wrote the court list class - there's one list in the public UI which still has court names hardcoded, and we want to bring them in from the list here, but we need them to be grouped slightly differently. This addess accessors to allow that!