Fix all court end dates to 2022
See [this trello card](https://trello.com/c/M3dSLGfX/263-change-year-ranges-for-browse-by-court) - Currently if no `end_year` is set for a court in the YAML, the displayed end_year is set to the current calendar year, which is misleading when we haven't yet ingested any judgments in the current year for that court. As a temporary fix, we've given every court a 2022 end date, which can be manually updated as and when we receive 2023 judgments for them. Longer term, this value should probably accessed dynamically from marklogic rather than set in static metadata.