Add lookup of court by code
In various places in the UI we need to return the full court name for a given internal court code.
This commit adds a method `courts.get_by_code()` which will return the court for a given code. We have also refactored to make sure that this method and the existing `courts.get_by_param()` run in constant time via a dictionary lookup, as this method will be used for every judgment returned in a results set, and we don't want to have to traverse the courts list each time.