Add a pre-commit check to validate courts data against a schema
At the moment there is nothing checking that the courts data is correct other than relying on a code review or hoping that a downstream test surfaces the problem. This defines what constitutes a valid `court_names.yaml` file, and checks it as part of pre-commit, so we can't commit invalid data.