Remove EWFC-B from list of court names and fold into EWFC.
The editor interface had an option for allowing editors to add EWFC-B as a court option.
This is incorrect. We in fact want it to appear nowhere, so deleting it from the list
of courts appears to be the correct answer.

It's not clear whether other non-listable, non-selectable courts should also be omitted
from the list of court names -- currently the editors can select any court. Some of these
are historic court names, KB/QB disambiguation etc.

EWFC-B still exists in the list of regular expressions that are NCNs in
src/ds_caselaw_utils/data/neutral_citation_regex.yaml .