refactor(types): move autogen court types to new types submodule
Utils will need to play host to multiple new types which explicitly describe concepts relating to courts; the existing autogenerated types from the schema should logically live in the same place, which is a new submobule.

This should not affect downstream consumers of the package, as these court types are only used internally.