feat(Court): make it clearer what Court.ncn actually means
`Court.ncn` is now `Court.ncn_pattern`, and is explicitly typed as being a subtype of a regular expression pattern. This is because it looked very similar to representing a Neutral Citation (ie the NCN of an individual judgment), and not (as it actually case) representing the pattern which valid NCNs for that court must match.

BREAKING CHANGE: `Court.ncn` is now `Court.ncn_pattern`, and now returns a `NeutralCitationPattern` which is a subtype of `re.Pattern[str]`.