feat: improve framing of the courts data schema
In readiness for storing more schemas in this repository, be more explicit with how we store the schema for courts data. This is now in a dedicated schemas folder, and we have added additional tooling to autogenerate documentation based on that schema.

BREAKING CHANGE: The location of the courts.schema.json file has changed, although the contents are identical.