build: change jsonschema-gentypes behaviour so config file is no longer schema-specific
Previously, the `jsonschema-gentypes.yaml` file was specific to a single schema. We now pass source and destination variables in directly, so we can reuse the s
ame configuration across multiple schemas. As part of this we retitle the root element in the courts schema to match the generated classes; this is a cosmetic ch
ange as far as the schema is concerned, not a functional change.